requirements.txt
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
fastapi==0.105.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
pytest==7.4.3
httpx==0.25.2 